Migrated init and loop code from Main class to the new Display class
